revert example and readme to basic example - add advanced example w/ protocol-in handlerName
